checkValues created on bot, and making logs better

+++ bot/index.js
@@ -3,15 +3,28 @@ require("dotenv").config();
 const Discord = require("discord.js");
 const axios = require("axios");
 
+const checkValues = require("./util/checkValues");
+
 const client = new Discord.Client();
 
 const prefix = "$";
 
 client.on("ready", () => {
-  console.log("Bot running");
+  console.log(`
+   _______________
+  |               |
+  |  Bot running  |
+  |_______________|
+
+  `);
 });
 
 client.on("guildCreate", async (guild) => {
+  if (!checkValues(process.env.BACKEND_LINK_DEV, "environment variable BACKEND_LINK_DEV")) return;
+  if (!checkValues(guild.id, "guild id")) return;
+  if (!checkValues(guild.name, "guild name")) return;
+  if (!checkValues(guild.ownerID, "guild owner id")) return;
+
   try {
     await axios.post(process.env.BACKEND_LINK_DEV + "/discord-server/add", {
       name: guild.name,
@@ -19,25 +32,40 @@ client.on("guildCreate", async (guild) => {
       discordOwnerId: guild.ownerID,
     });
 
-    console.log("Discord server added");
+    console.log(`\n\tDiscord server with name: ${guild.name} and id: ${guild.id} added\n`);
     return;
   } catch (err) {
-    console.error("Error on add discord server: ", err);
+    console.error(
+      `\n\tError on add discord server with name: ${guild.name} and id: ${guild.id} \n`
+    );
+    console.error(err);
     return;
   }
 });
 
 client.on("guildDelete", async (guild) => {
+  if (!checkValues(process.env.BACKEND_LINK_DEV, "environment variable BACKEND_LINK_DEV")) return;
+  if (!checkValues(guild.id, "guild id")) return;
+  if (!checkValues(guild.name, "guild name")) return;
+
   try {
     await axios.delete(process.env.BACKEND_LINK_DEV + "/discord-server/" + guild.id + "/remove");
 
-    return console.log("Discord server deleted");
+    console.log(`\n\tDiscord server with name: ${guild.name} and id: ${guild.id} deleted\n`);
+    return;
   } catch (err) {
-    return console.error("Error on delete discord server: ", err);
+    console.error(
+      `\n\tError on delete discord server with name: ${guild.name} and id: ${guild.id}\n`
+    );
+    console.error(err);
+    return;
   }
 });
 
 client.on("message", (message) => {
+  if (!checkValues(message.content, "message content")) return;
+  if (!checkValues(message.channel, "message channel")) return;
+
   if (message.content === prefix + "ping") {
     message.channel.send("pong");
   }
